Add two Tasker profiles which respond on NMA notifications.
The two profiles respond to the "User is idle/active" notifications sent
via NMA to the Android phone. Tasker will turn on/off the synchronization.

Notify My Android needs to be installed in order to make these Tasker
profiles work.

The two profiles are connected through the variable IGNOREIDLE, which
makes sure that the synchronization is not turned back on when you are
simply switching computers, both running Activity2Phone.