Remove hack'ish setTimeout for calling setActive after loadSettings.
loadSettings performs async operations, so setActive can only be called
when all settings are loaded (to be more specific, the API key). Before
this was done rather ugly with a 1s timeout. Now the setActive function
is passed as a callback function to loadSettings, such that it will be
called at the exact moment when all data is loaded.